Change all namespaces
I change all namespaces to reduce confusion between the framework core and the framework project.
