--- src/Exceptions/BaseException.php
@@ -1,15 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Theozebua\OzeFramework\Exceptions;
-
-use Throwable;
-
-/**
- * This is the base exception interface of Oze Framework.
- */
-interface BaseException extends Throwable
-{
-    // 
-}
