--- src/Exceptions/Controller/ControllerNotFoundException.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Theozebua\OzeFramework\Exceptions\Controller;
-
-use Exception;
-use Theozebua\OzeFramework\Exceptions\BaseException;
-
-final class ControllerNotFoundException extends Exception implements BaseException
-{
-    // 
-}
