--- src/Exceptions/Request/KeyNotFoundException.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Theozebua\OzeFramework\Exceptions\Request;
-
-use Exception;
-use Theozebua\OzeFramework\Exceptions\BaseException;
-
-final class KeyNotFoundException extends Exception implements BaseException
-{
-    // 
-}
