--- src/Exceptions/Router/RouteNotFoundException.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Theozebua\OzeFramework\Exceptions\Router;
-
-use Exception;
-use Theozebua\OzeFramework\Exceptions\BaseException;
-
-final class RouteNotFoundException extends Exception implements BaseException
-{
-    // 
-}
