--- src/Exceptions/View/ViewNotFoundException.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Theozebua\OzeFramework\Exceptions\View;
-
-use Exception;
-use Theozebua\OzeFramework\Exceptions\BaseException;
-
-final class ViewNotFoundException extends Exception implements BaseException
-{
-    // 
-}
