+++ src/OzeFramework/Exceptions/BaseException.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OzeFramework\Exceptions;
+
+use Throwable;
+
+/**
+ * This is the base exception interface of Oze Framework.
+ */
+interface BaseException extends Throwable
+{
+    // 
+}
