+++ src/OzeFramework/Exceptions/Container/NotFoundException.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OzeFramework\Exceptions\Container;
+
+use Exception;
+use Psr\Container\NotFoundExceptionInterface;
+use OzeFramework\Exceptions\BaseException;
+
+final class NotFoundException extends Exception implements NotFoundExceptionInterface, BaseException
+{
+    // 
+}
