+++ src/OzeFramework/Exceptions/Controller/ControllerNotFoundException.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OzeFramework\Exceptions\Controller;
+
+use Exception;
+use OzeFramework\Exceptions\BaseException;
+
+final class ControllerNotFoundException extends Exception implements BaseException
+{
+    // 
+}
