+++ src/OzeFramework/Exceptions/Request/KeyNotFoundException.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OzeFramework\Exceptions\Request;
+
+use Exception;
+use OzeFramework\Exceptions\BaseException;
+
+final class KeyNotFoundException extends Exception implements BaseException
+{
+    // 
+}
