+++ src/OzeFramework/Exceptions/Router/RouteNotFoundException.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OzeFramework\Exceptions\Router;
+
+use Exception;
+use OzeFramework\Exceptions\BaseException;
+
+final class RouteNotFoundException extends Exception implements BaseException
+{
+    // 
+}
