+++ src/OzeFramework/Exceptions/View/ViewNotFoundException.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OzeFramework\Exceptions\View;
+
+use Exception;
+use OzeFramework\Exceptions\BaseException;
+
+final class ViewNotFoundException extends Exception implements BaseException
+{
+    // 
+}
